src/segmentation.py
"""
Segmentation module for Interior Style Transfer POC.

This module handles image segmentation using SAM (Segment Anything Model)
to create masks for preserving structural elements in interiors.
"""

import os
import numpy as np
from typing import Dict, List, Tuple, Any, Optional
from PIL import Image
import cv2

# Stub for SAM import - will be implemented in Story 3
# from segment_anything import SamPredictor, SamAutomaticMaskGenerator, sam_model_registry


class SegmentationHandler:
    """
    Handles segmentation of interior images to identify structural elements.
    
    Uses Segment Anything Model (SAM) to create masks for walls, floors,
    ceilings, and other architectural elements.
    """
    
    def __init__(self, model_type: str = "vit_h", checkpoint_path: Optional[str] = None):
        """
        Initialize the segmentation handler with SAM model.
        
        Args:
            model_type: SAM model type ('vit_h', 'vit_l', 'vit_b')
            checkpoint_path: Path to model weights (or None to download)
        """
        self.model_type = model_type
        self.checkpoint_path = checkpoint_path
        self.device = "cuda" if self._is_cuda_available() else "cpu"
        self.model = None
        self.predictor = None
        self.mask_generator = None
        
        # Placeholder for future implementation
        print("SegmentationHandler initialized (stub for future implementation)")
    
    def _is_cuda_available(self) -> bool:
        """Check if CUDA is available for GPU acceleration."""
        try:
            import torch
            return torch.cuda.is_available()
        except ImportError:
            return False
    
    def load_model(self) -> bool:
        """
        Load the SAM model with the specified checkpoint.
        
        Returns:
            bool: True if model loaded successfully
        """
        # Placeholder for future implementation (Story 3)
        return True
    
    def generate_masks(self, image: np.ndarray) -> Dict[str, np.ndarray]:
        """
        Generate segmentation masks for an interior image.
        
        Args:
            image: Input image as NumPy array
            
        Returns:
            Dict mapping mask types to binary masks
        """
        # Placeholder for future implementation (Story 3)
        h, w = image.shape[:2]
        placeholder_mask = np.zeros((h, w), dtype=np.uint8)
        
        return {
            "walls": placeholder_mask,
            "floor": placeholder_mask,
            "ceiling": placeholder_mask,
            "windows": placeholder_mask,
            "structure": placeholder_mask  # Combined structural elements
        }
    
    def apply_mask(self, image: np.ndarray, mask: np.ndarray, 
                  invert: bool = False) -> np.ndarray:
        """
        Apply a binary mask to an image.
        
        Args:
            image: Input image as NumPy array
            mask: Binary mask as NumPy array
            invert: Whether to invert the mask
            
        Returns:
            Masked image as NumPy array
        """
        if invert:
            mask = 1 - mask
            
        # Convert mask to 3-channel if needed
        if len(mask.shape) == 2:
            mask = np.repeat(mask[:, :, np.newaxis], 3, axis=2)
        
        return image * mask
