requirements.txt
# Core dependencies
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
pillow>=9.4.0
opencv-python>=4.7.0
scikit-image>=0.20.0
ipywidgets>=8.0.0

# API integrations
fal>=0.10.0
segment-anything>=1.0  # Meta's SAM

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0

# Notebook
jupyter>=1.0.0
notebook>=6.5.0
